# backend/docs/swagger.yaml
openapi: 3.0.0
info: 
  title: SpotPlay
  description: 'API platform to show music info...'
  contact:
    email: support@spotplay.com
  license: 
    name: Apache 2.0
    url: http://apac...
  version: 1.0.0
servers:
- url: http://localhost:3000/api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      description: JSON authorization header using a token
      scheme: bearer
      bearerFormat: JWT 
  schemas:
...
    Song:
      type: object
      properties:
        title:
          type: string
        uri:
          type: string
        duration:
          type: string
        year:
          type: integer
          format: int64
        imagen:
          type: string
        artistId:
          type: string
        genreId:
          type: string
        albumId:
          type: string
    User:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    List:
      type: object
      properties:
        name:
          type: string
        userId:
          type: string
...
security:
  - bearerAuth: []
paths:
...
  /user:
    get:
      tags:
      - User
      summary: Get all users
      description: Get al users registered into the app
      responses:
        200:
          description: "Ok: Users info returned"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
  /user/signup:
    post:
      tags:
      - User
      summary: Create a new user
      operationId: createUser
      description: Method to create a new user that will be able to use the app
      requestBody:
        description: Send user data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        201:
          description: "Created: Resurce was created successfully"
        400:
          description: "Bad request: Bad format or parameter missing"
  /user/{id}:
    get:
      tags:
      - User
      summary: Get specific user by ID
      description: Get a user by ID
      parameters:
      - name: id
        in: path
        description: ID for the user to return
        required: true
        type: string
      responses:
        200:
          description: "Ok: User info returned"
        404:
          description: "Not found: Resource required not found"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
    delete:
      tags:
      - User
      summary: Delete specific user by ID
      description: Delete a user by ID
      parameters:
      - name: id
        in: path
        description: ID for the user to delete
        required: true
        type: string
      responses:
        200:
          description: "Ok: User was deleted"
        404:
          description: "Not found: Resource required not found"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
    put:
      tags:
      - User
      summary: Update specific user by ID
      description: Update a user by ID
      parameters:
      - name: id
        in: path
        description: ID for the user to delete
        required: true
        type: string
      requestBody:
        description: Send new user data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true        
      responses:
        200:
          description: "Ok: User was updated successfully"
        400:
          description: "Bad request: Bad format or parameter missing"
        404:
          description: "Not found: Resource required not found"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
  /song:
    get:
      tags:
      - Song
      summary: Get all songs
      description: Get al songs registered into the app.
      responses:
        200:
          description: "Ok: Songs info returned"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
    post:
      tags:
      - Song
      summary: Create a new song
      operationId: createSong
      description: Method to create a new song
      requestBody:
        description: Send song data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Song'
        required: true
      responses:
        201:
          description: "Created: Resurce was created successfully"
        400:
          description: "Bad request: Bad format or parameter missing"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
  /song/{id}:
    get:
      tags:
      - Song
      summary: Get specific song by ID
      description: Get a song by ID
      parameters:
      - name: id
        in: path
        description: ID for the song to return
        required: true
        type: string
      responses:
        200:
          description: "Ok: Song info returned"
        404:
          description: "Not found: Resource required not found"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
    delete:
      tags:
      - Song
      summary: Delete specific song by ID
      description: Delete a song by ID
      parameters:
      - name: id
        in: path
        description: ID for the song to delete
        required: true
        type: string
      responses:
        200:
          description: "Ok: Song was deleted"
        404:
          description: "Not found: Resource required not found"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
    put:
      tags:
      - Song
      summary: Update specific song by ID
      description: Update a song by ID
      parameters:
      - name: id
        in: path
        description: ID for the song to delete
        required: true
        type: string
      requestBody:
        description: Send new song data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Song'
        required: true        
      responses:
        200:
          description: "Ok: Song was updated successfully"
        400:
          description: "Bad request: Bad format or parameter missing"
        404:
          description: "Not found: Resource required not found"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
  /list:
    get:
      tags:
      - PlayList
      summary: Get all lists
      description: Get al lists registered into the app.
      responses:
        200:
          description: "Ok: List info returned"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
    post:
      tags:
      - PlayList
      summary: Create a new list
      operationId: createList
      description: Method to create a new list
      requestBody:
        description: Send list data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/List'
        required: true
      responses:
        201:
          description: "Created: Resurce was created successfully"
        400:
          description: "Bad request: Bad format or parameter missing"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
  /list/{id}:
    get:
      tags:
      - PlayList
      summary: Get specific list by ID
      description: Get a list by ID
      parameters:
      - name: id
        in: path
        description: ID for the list to return
        required: true
        type: string
      responses:
        200:
          description: "Ok: List info returned"
        404:
          description: "Not found: Resource required not found"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
    delete:
      tags:
      - PlayList
      summary: Delete specific list by ID
      description: Delete a list by ID
      parameters:
      - name: id
        in: path
        description: ID for the list to delete
        required: true
        type: string
      responses:
        200:
          description: "Ok: List was deleted"
        404:
          description: "Not found: Resource required not found"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
    put:
      tags:
      - PlayList
      summary: Update specific list by ID
      description: Update a list by ID
      parameters:
      - name: id
        in: path
        description: ID for the list to delete
        required: true
        type: string
      requestBody:
        description: Send new list data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/List'
        required: true        
      responses:
        200:
          description: "Ok: List was updated successfully"
        400:
          description: "Bad request: Bad format or parameter missing"
        404:
          description: "Not found: Resource required not found"
        401:
          description: "Unauthorized: Not enough authorization to perform the action"
...